Separate BpTree handling functions into BpTreeNode
Separation of concerns
