+++ src/realm/column_binary.hpp
@@ -108,9 +108,9 @@ private:
 
     // Called by Array::bptree_insert().
     static ref_type leaf_insert(MemRef leaf_mem, ArrayParent&, size_t ndx_in_parent, Allocator&, size_t insert_ndx,
-                                Array::TreeInsert<BinaryColumn>& state);
+                                BpTreeNode::TreeInsert<BinaryColumn>& state);
 
-    struct InsertState : Array::TreeInsert<BinaryColumn> {
+    struct InsertState : BpTreeNode::TreeInsert<BinaryColumn> {
         bool m_add_zero_term;
     };
 
@@ -130,7 +130,7 @@ private:
 
     void leaf_to_dot(MemRef, ArrayParent*, size_t ndx_in_parent, std::ostream&) const override;
 
-    friend class Array;
+    friend class BpTreeNode;
     friend class ColumnBase;
 };
 
@@ -200,7 +200,7 @@ inline size_t BinaryColumn::size() const noexcept
         return leaf->size();
     }
     // Non-leaf root
-    return m_array->get_bptree_size();
+    return static_cast<BpTreeNode*>(m_array.get())->get_bptree_size();
 }
 
 inline bool BinaryColumn::is_nullable() const noexcept
@@ -251,7 +251,7 @@ inline BinaryData BinaryColumn::get(size_t ndx) const noexcept
     }
 
     // Non-leaf root
-    std::pair<MemRef, size_t> p = m_array->get_bptree_leaf(ndx);
+    std::pair<MemRef, size_t> p = static_cast<BpTreeNode*>(m_array.get())->get_bptree_leaf(ndx);
     const char* leaf_header = p.first.get_addr();
     size_t ndx_in_leaf = p.second;
     Allocator& alloc = m_array->get_alloc();
@@ -420,7 +420,7 @@ inline size_t BinaryColumn::get_size_from_ref(ref_type root_ref, Allocator& allo
         // Big blobs leaf
         return ArrayBigBlobs::get_size_from_header(root_header);
     }
-    return Array::get_bptree_size_from_header(root_header);
+    return BpTreeNode::get_bptree_size_from_header(root_header);
 }
