+++ src/realm/column_linklist.cpp
@@ -211,7 +211,7 @@ void LinkListColumn::cascade_break_backlinks_to(size_t row_ndx, CascadeState& st
     ref_type ref = get_as_ref(row_ndx);
     if (ref == 0)
         return;
-    Array root(get_alloc());
+    BpTreeNode root(get_alloc());
     root.init_from_ref(ref);
 
     if (!root.is_inner_bptree_node()) {
@@ -267,7 +267,7 @@ void LinkListColumn::cascade_break_backlinks_to_all_rows(size_t num_rows, Cascad
 
     // Avoid the construction of both a LinkView and a IntegerColumn instance,
     // since both would involve heap allocations.
-    Array root(get_alloc()), leaf(get_alloc());
+    BpTreeNode root(get_alloc()), leaf(get_alloc());
     for (size_t i = 0; i < num_rows; ++i) {
         ref_type ref = get_as_ref(i);
         if (ref == 0)
@@ -788,8 +788,7 @@ void LinkListColumn::verify(const Table& table, size_t col_ndx) const
 
 std::pair<ref_type, size_t> LinkListColumn::get_to_dot_parent(size_t ndx_in_parent) const
 {
-    std::pair<MemRef, size_t> p = get_root_array()->get_bptree_leaf(ndx_in_parent);
-    return std::make_pair(p.first.get_ref(), p.second);
+    return IntegerColumn::get_to_dot_parent(ndx_in_parent);
 }
 
 // LCOV_EXCL_STOP ignore debug functions
